requirements.txt
tensorflow
numpy
pandas
opencv-python
matplotlib
scikit-learn
pillow
tqdm
streamlit
fastapi
uvicorn[standard]
python-multipart
efficientnet
seaborn
